Fixed compiler warnings.

+++ conex/test/exponential_map_test.cc
@@ -58,7 +58,7 @@ class TestCases : public testing::Test {
 
     auto result = T::Zero(n, n);
     ExponentialMap(arg, &result);
-    for (int k = 0; k < result.size(); k++) {
+    for (size_t k = 0; k < result.size(); k++) {
       for (int i = 0; i < n; i++) {
         for (int j = 0; j < n; j++) {
           EXPECT_NEAR(reference.at(k)(i, j), result.at(k)(i, j), 1e-7);
